Augmented-data projection: Internals

The augmented-data projection makes extensive use of augmented-rows matrices and augmented-length vectors. In the following, \(N\), \(C_{\mathrm{cat}}\), \(C_{\mathrm{lat}}\), \(S_{\mathrm{ref}}\), and \(S_{\mathrm{prj}}\) from help topic refmodel-init-get are used. Furthermore, let \(C\) denote either \(C_{\mathrm{cat}}\) or \(C_{\mathrm{lat}}\), whichever is appropriate in the context where it is used (e.g., for ref_predfun's output, \(C = C_{\mathrm{lat}}\)). Similarly, let \(S\) denote either \(S_{\mathrm{ref}}\) or \(S_{\mathrm{prj}}\), whichever is appropriate in the context where it is used. Then an augmented-rows matrix is a matrix with \(N \cdot C\) rows in \(C\) blocks of \(N\) rows, i.e., with the \(N\) observations nested in the \(C\) (possibly latent) response categories. For ordered response categories, the \(C\) (possibly latent) response categories (i.e., the row blocks) have to be sorted increasingly. The columns of an augmented-rows matrix have to correspond to the \(S\) parameter draws, just like for the traditional projection. An augmented-rows matrix is of class augmat (inheriting from classes matrix and array) and needs to have the value of \(C\) stored in an attribute called ndiscrete. An augmented-length vector (class augvec) is the vector resulting from subsetting an augmented-rows matrix to extract a single column and thereby dropping dimensions.