Cross-validate the variable selection (varsel)

Perform cross-validation for the projective variable selection for a generalized linear model.

cv_varsel(fit, method = NULL, cv_method = NULL, ns = NULL, nc = NULL,
  nspred = NULL, ncpred = NULL, nv_max = NULL, intercept = NULL,
  penalty = NULL, verbose = T, nloo = 100, K = NULL, k_fold = NULL,
  lambda_min_ratio = 1e-05, nlambda = 500, regul = 1e-06,
  validate_search = T, seed = NULL, ...)

Arguments

fit
Same as in varsel.

method
Same as in varsel.

cv_method
The cross-validation method, either 'LOO' or 'kfold'. Default is 'LOO'.

ns
Number of samples used for selection. Ignored if nc is provided or if method='L1'.

nc
Number of clusters used for selection. Default is 1 and ignored if method='L1' (L1-search uses always one cluster).

nspred
Number of samples used for prediction (after selection). Ignored if ncpred is given.

ncpred
Number of clusters used for prediction (after selection). Default is 5.

nv_max
Same as in varsel.

intercept
Same as in varsel.

penalty
Same as in varsel.

verbose
Whether to print out some information during the validation, Default is TRUE.

nloo
Number of observations used to compute the LOO validation (anything between 1 and the total number of observations). Smaller values lead to faster computation but higher uncertainty (larger errorbars) in the accuracy estimation. Default value is 100. Only applicable if cv_method = LOO.

K
Number of folds in the k-fold cross validation. Only applicable if cv_method = TRUE and k_fold = NULL.

k_fold
An array with cross-validated stanfits and the respective test datasets returned by kfold with save_fits=TRUE. If not provided, kfold is called inside the function.

lambda_min_ratio
Same as in varsel.

nlambda
Same as in varsel.

regul
Amount of regularization in the projection. Usually there is no need for regularization, but sometimes for some models the projection can be ill-behaved and we need to add some regularization to avoid numerical problems. Default is 1e-9.

validate_search
Whether to cross-validate also the selection process, that is, whether to perform selection separately for each fold. Default is TRUE and we strongly recommend not setting this to FALSE, because this is known to bias the accuracy estimates for the selected submodels. However, setting this to FALSE can sometimes be useful because comparing the results to the case where this parameter is TRUE gives idea how strongly the feature selection is (over)fitted to the data (the difference corresponds to the search degrees of freedom or the effective number of parameters introduced by the selectin process).

seed
Random seed used in the subsampling LOO. By default uses a fixed seed.

...
Currently ignored.

Value

The original stanreg object augmented with an element 'varsel', which is a list containing the following elements:

vind

The order in which the variables were added to the submodel.

pctch

Percentage of cross-validation runs that included the given variable to a model of given size.

kl

KL-divergence for each submodel size.

summaries

CV-summary statistics computed during the selection.

d_test

The data used to evaluate the summaries.

family_kl

A modified family-object.

Examples

### Usage with stanreg objects
fit <- stan_glm(y~x, binomial())

#> Warning: Omitting the 'data' argument is not recommended and may not be allowed in future versions of rstanarm. Some post-estimation functions (in particular 'update', 'loo', 'kfold') are not guaranteed to work properly unless 'data' is specified as a data frame.

#> Error in eval(predvars, data, env): object 'y' not found

fit_cv <- cv_varsel(fit)

#> Warning: Some Pareto k diagnostic values are slightly high. See help('pareto-k-diagnostic') for details.

#> [1] "Computing LOOs..."
#>   |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
#> [1] "Performing the selection using all the data.."
#> [1] "Done."