Generic reference model initialization

Initializes a structure that can be used as a reference fit for the projective variable selection. This function is provided to allow construction of the reference fit using also other tools than rstanarm, because only certain specific information is needed for the actual projection and variable selection.

init_refmodel(x, y, family, predfun = NULL, dis = NULL, offset = NULL,
  wobs = NULL, wsample = NULL, intercept = TRUE, cvfits = NULL)

Arguments

x
Predictor matrix of dimension n-by-D containing the candidate variables for selection (i.e. variables from which to select the submodel). Rows denote the observations and columns the different variables.

y
Vector of length n giving the target variable values.

family
family object giving the model family

predfun
Function that takes a nt-by-D test predictor matrix as an input (nt = # test points, D = # predictors) and outputs a nt-by-S matrix of expected values for the target variable y, each column corresponding to one posterior draw for the parameters in the reference model (the number of draws S can also be 1). The output should be computed without any offsets, these are automatically taken into account internally, e.g. in cross-validation.

dis
Vector of length S giving the posterior draws for the dispersion parameter in the reference model if there is such a parameter in the model family. For Gaussian observation model this is the noise std sigma.

offset
Offset to be added to the linear predictor in the projection. (Same as in function glm.)

wobs
Observation weights. If omitted, equal weights are assumed.

wsample
vector of length S giving the weights for the posterior draws. If omitted, equal weights are assumed.

intercept
Whether to use intercept. Default is TRUE.

cvfits
A list with K elements, each of which is a list with fields including at least variables: tr, ts and predfun giving the training and test indices and prediction function for each fold. Additionally each element can have field dis (dispersion samples for each fold) if the model has a dispersion parameter. Can be omitted but needed for K-fold cross validation for genuine reference models.

Value

An object that can be passed to all the functions that take the reference fit as the first argument, such as varsel, cv_varsel, proj_predict and proj_linpred.