Projection predictive feature selection
Source: R/projpred-package.R

Description

projpred is an R package to perform projection predictive variable selection for generalized linear models. The package is aimed to be compatible with rstanarm but also other reference models can be used (see function init_refmodel).

Currently, the supported models (family objects in R) include Gaussian, Binomial and Poisson families, but more will be implemented later. See the quickstart-vignette for examples.

Functions

varsel, cv_varsel, init_refmodel

Perform and cross-validate the variable selection. init_refmodel can be used to initialize a reference model other than rstanarm-fit.

project

Get the projected posteriors of the reduced models.

proj_predict, proj_linpred

Make predictions with reduced number of features.

varsel_plot, varsel_stats

Visualize and get some key statistics about the variable selection.

References

Dupuis, J. A. and Robert, C. P. (2003). Variable selection in qualitative models via an entropic explanatory power. Journal of Statistical Planning and Inference, 111(1-2):77–94.

Goutis, C. and Robert, C. P. (1998). Model choice in generalised linear models: a Bayesian approach via Kullback–Leibler projections. Biometrika, 85(1):29–37.

Juho Piironen and Aki Vehtari (2017). Comparison of Bayesian predictive methods for model selection. Statistics and Computing, 27(3):711-735. doi:10.1007/s11222-016-9649-y. (Online).