Suggest model size

This function can be used for suggesting an appropriate model size based on certain rule. Notice that the decision rules are heuristic and should be interpreted as guidelines. It is recommended that the user studies the results via varsel_plot and or varsel_stats and makes the final decision based on what is most appropriate for the given problem.

suggest_size(object, stat = "elpd", alpha = 0.32, pct = 0,
  type = "upper", warnings = TRUE, ...)

Arguments

object
The object returned by varsel or cv_varsel.

stat
Statistic used for the decision. Default is elpd. See varsel_stats for other possible choices.

alpha
A number indicating the desired coverage of the credible intervals based on which the decision is made. E.g. alpha=0.1 corresponds to 90% probability mass within the intervals. See details for more information.

pct
Number indicating the relative proportion between full model and null model utilities one is willing to sacrifice. See details for more information.

type
Either 'upper' (default) or 'lower' determining whether the decisions are based on the upper or lower credible bounds. See details for more information.

warnings
Whether to give warnings if automatic suggestion fails, mainly for internal use. Default is TRUE, and usually no reason to set to FALSE.

...
Currently ignored.

Details

The suggested model size is the smallest model for which either the lower or upper (depending on argument type) credible bound of the submodel utility \(u_k\) with significance level alpha falls above $$u_ref - pct*(u_ref - u_0)$$ Here \(u_ref\) denotes the reference model utility and \(u_0\) the null model utility (currently the utility is taken to be the mean log predictive density, MLPD). The lower and upper bounds are defined to contain the submodel utility with probability 1-alpha (each tail has mass alpha/2).

By default ratio=0, alpha=0.32 and type='upper' which means that we select the smallest model for which the upper tail exceeds the reference model level, that is, which is better than the reference model with probability 0.16 (and consequently, worse with probability 0.84). In other words, the estimated difference between the reference model and submodel utitlities is at most one standard error away from zero, so the two utilities are considered to be close.